pyproject.toml
[project]
name = "pulsemaps-backend"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "pydantic-settings",
  "python-dateutil",
  "httpx",
  "langchain",
  "langchain-openai",
  "langgraph",
]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP"]
